drm/msm/sde: Fix incompatible pointer type warning
Clang warns:

../drivers/gpu/drm/msm/sde/sde_plane.c:5294:24: warning: incompatible pointer types passing 'unsigned long *' to parameter of type 'u32 *' (aka 'unsigned int *') [-Wincompatible-pointer-types]
                        psde->debugfs_root, &psde->features);
                                            ^~~~~~~~~~~~~~~
../include/linux/debugfs.h:118:35: note: passing argument to parameter 'value' here
                                  struct dentry *parent, u32 *value);
                                                              ^
1 warning generated.

Switch the features field to uint32_t to match the state of this driver
in mainline.

+++ drivers/gpu/drm/msm/sde/sde_plane.c
@@ -106,7 +106,7 @@ struct sde_plane {
 	struct mutex lock;
 
 	enum sde_sspp pipe;
-	unsigned long features;      /* capabilities from catalog */
+	uint32_t features;      /* capabilities from catalog */
 	uint32_t nformats;
 	uint32_t formats[64];
